requirements.txt
streamlit==1.36.0
openai==1.35.15
pillow==10.2.0